Refactor Default Apps into same file as Edit Modal component
